requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
Faker==33.1.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
